float to double

+++ boundary.cpp
@@ -1,25 +1,25 @@
 #include "boundary.h"
 
-void boundarypsi(float *psi, int m, int n, int b, int h, int w) {
+void boundarypsi(double *psi, int m, int n, int b, int h, int w) {
     int i, j;
 
     // BCs on bottom edge
 
     for (i = b + 1; i <= b + w - 1; i++) {
-        psi[i * (m + 2) + 0] = (float) (i - b);
+        psi[i * (m + 2) + 0] = (double) (i - b);
     }
 
     for (i = b + w; i <= m; i++) {
-        psi[i * (m + 2) + 0] = (float) (w);
+        psi[i * (m + 2) + 0] = (double) (w);
     }
 
     // BCs on RHS
 
     for (j = 1; j <= h; j++) {
-        psi[(m + 1) * (m + 2) + j] = (float) w;
+        psi[(m + 1) * (m + 2) + j] = (double) w;
     }
 
     for (j = h + 1; j <= h + w - 1; j++) {
-        psi[(m + 1) * (m + 2) + j] = (float) (w - j + h);
+        psi[(m + 1) * (m + 2) + j] = (double) (w - j + h);
     }
 }
